backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.129.0",
    "openai>=2.21.0",
    "openai-agents>=0.9.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
